Move the Test Connection String Endpoint to the Administration Controller
