+++ src/SimpleSocial/Web/SimpleSocial.Web/Controllers/HomeController.cs
@@ -1,18 +1,11 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Configuration;
-using System;
 
 namespace SimpleSocial.Web.Controllers
 {
     public class HomeController : BaseController
     {
-        private readonly IConfiguration configuration;
 
-        public HomeController(IConfiguration configuration)
-        {
-            this.configuration = configuration;
-        }
         public IActionResult Index()
         {
             if (User.Identity.IsAuthenticated)
@@ -33,14 +26,6 @@ namespace SimpleSocial.Web.Controllers
         public IActionResult Error()
         {
             return View();
-        }
-
-        public IActionResult ConnectionString()
-        {
-            var envName = Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT");
-            var conString = this.configuration.GetConnectionString("DefaultConnection");
-            var result = envName + Environment.NewLine + conString;
-            return this.Json(result);
         }
     }
 }
